CircuitPython Controller: fix to work on Trinket M0.
Trinket M0 does not support `time.time()`, but does support
`time.monotonic()`, which is also in seconds and works fine for the need of
the program. But on the other hand, `monotonic()` returns a floating point
value.

Also correct the detection of the MCP23017, and raise an exception if no
expander is found at all.