Clean up the Trinket M0 based boards.
This makes the two board use the same size, and the same positioning of the common components.

It also reorients the silkscreen markings on the back to have the same orientation as the actuator board.

Also mark the used lines from the Trinket, as you can technically use any other MCU that can drive I²C.