Add an ATmega48 based firmware.
This is based on the board designs in eagle-boards/atmega/ and the original
8051 firmware.

Unlike the STC89C52, the ATmega48 MCU has non-linear mapping of the I/Os,
which makes the room definition a bit more complicated, but otherwise the
logic of the code is essentially the same.

Note that the "strobe" test has been removed from this version, because
it's unsafe: it would connect the ground of multiple LEDs on the same
Darlington collector in the real setup.